Added vouchers and content instance versions to sdk, restructured pub methods
