+++ lib/pub.rb
@@ -3,6 +3,9 @@
 require 'yaml'
 require_relative './client'
 require_relative './mints/helpers/mints_helper'
+require_relative './pub/content/content'
+require_relative './pub/ecommerce/ecommerce'
+require_relative './pub/config/config'
 
 module Mints
   ##
@@ -61,6 +64,9 @@ module Mints
     attr_reader :client
 
     include MintsHelper
+    include Content
+    include Ecommerce
+    include Config
 
     ##
     # === Initialize.
@@ -116,433 +122,6 @@ module Mints
     #     @data = @mints_pub.register_visit_timer("60da2325d29acc7e55684472", 4)
     def register_visit_timer(visit, time)
       @client.raw('get', "/register-visit-timer?visit=#{visit}&time=#{time}")
-    end
-
-    ### V1/CONTENT ###
-
-    ##
-    # === Get Asset Info.
-    # Get a description of an Asset.
-    #
-    # ==== Parameters
-    # slug:: (String) -- It's the string identifier of the asset.
-    #
-    # ==== Example
-    #     @data = @mints_pub.get_asset_info("asset_slug")
-    def get_asset_info(slug)
-      @client.raw('get', "/content/asset-info/#{slug}")
-    end
-
-    ##
-    # === Get Stories.
-    # Get a collection of stories.
-    #
-    # ==== Parameters
-    # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
-    # use_post:: (Boolean) -- Variable to determine if the request is by 'post' or 'get' functions.
-    #
-    # ==== First Example
-    #     @data = @mints_pub.get_stories
-    #
-    # ==== Second Example
-    #     options = { fields: 'id, slug' }
-    #     @data = @mints_pub.get_stories(options)
-    #
-    # ==== Third Example
-    #     options = {
-    #       fields: 'id, slug'
-    #     }
-    #     @data = @mints_pub.get_stories(options, false)
-    def get_stories(options = nil, use_post = true)
-      get_query_results('/content/stories', options, use_post)
-    end
-
-    ##
-    # === Get Story.
-    # Get a single story.
-    #
-    # ==== Parameters
-    # slug:: (String) -- It's the string identifier generated by Mints.
-    # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
-    #
-    # ==== First Example
-    #     @data = @mints_pub.get_story("story_slug")
-    #
-    # ==== Second Example
-    #     @data = @mints_pub.get_story("story_slug", options.to_json)
-    def get_story(slug, options = nil)
-      @client.raw('get', "/content/stories/#{slug}", options)
-    end
-
-    ##
-    # === Get Story Versions.
-    # Get a collection of story version.
-    #
-    # ==== Parameters
-    # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
-    # use_post:: (Boolean) -- Variable to determine if the request is by 'post' or 'get' functions.
-    #
-    # ==== First Example
-    #     @data = @mints_pub.get_story_versions
-    #
-    # ==== Second Example
-    #     options = {
-    #       fields: 'id, title'
-    #     }
-    #     @data = @mints_pub.get_story_versions(options)
-    #
-    # ==== Third Example
-    #     options = {
-    #       fields: 'id, title'
-    #     }
-    #     @data = @mints_pub.get_story_versions(options, false)
-    def get_story_versions(options = nil, use_post = true)
-      get_query_results('/content/story-versions', options, use_post)
-    end
-
-    ##
-    # === Get Story Version.
-    # Get a single story version.
-    #
-    # ==== Parameters
-    # slug:: (String) -- It's the string identifier generated by Mints.
-    # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
-    #
-    # ==== First Example
-    #     @data = @mints_pub.get_story_version("story_slug")
-    #
-    # ==== Second Example
-    #     @data = @mints_pub.get_story_version("story_slug", options.to_json)
-    def get_story_version(slug, options = nil)
-      @client.raw('get', "/content/story-versions/#{slug}", options)
-    end
-
-    ##
-    # === Get Forms.
-    # Get a collection of forms.
-    #
-    # ==== Example
-    #     @data = @mints_pub.get_forms
-    def get_forms(options = nil)
-      @client.raw('get', '/content/forms', options)
-    end
-
-    ##
-    # === Get Form.
-    # Get a single form.
-    #
-    # ==== Parameters
-    # slug:: (String) -- It's the string identifier generated by Mints.
-    #
-    # ==== Example
-    #     @data = @mints_pub.get_form("form_slug")
-    def get_form(slug, options = nil)
-      @client.raw('get', "/content/forms/#{slug}", options)
-    end
-
-    ##
-    # === Submit Form.
-    # Submit a form with data.
-    #
-    # ==== Parameters
-    # data:: (Hash) -- Data to be submitted.
-    #
-    # ==== Example
-    #     data = {
-    #       'form_slug': 'form_slug',
-    #       'email': 'email@example.com',
-    #       'given_name': 'given_name',
-    #       'f1': 'Field 1 answer',
-    #       'f2': 'Field 2 answer',
-    #       'f3': 'Field 3 answer'
-    #     }
-    #     @data = @mints_pub.submit_form(data)
-    def submit_form(data)
-      @client.raw('post', '/content/forms/submit', nil, data_transform(data))
-    end
-
-    ##
-    # === Get Content Instances.
-    # Get a collection of content instances. _Note:_ Options must be specified.
-    #
-    # ==== Parameters
-    # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
-    #
-    # ==== First Example
-    #     options = {
-    #       "template": "content_instance_template_slug"
-    #     }
-    #     @data = @mints_pub.get_content_instances(options)
-    #
-    # ==== Second Example
-    #     options = {
-    #       "template": "content_instance_template_slug",
-    #       sort: "-id"
-    #     }
-    #     @data = @mints_pub.get_content_instances(options)
-    def get_content_instances(options = nil)
-      @client.raw('get', '/content/content-instances', options)
-    end
-
-    ##
-    # === Get Content Instance.
-    # Get a single content instance.
-    #
-    # ==== Parameters
-    # slug:: (String) -- It's the string identifier generated by Mints.
-    #
-    # ==== Example
-    #     @data = @mints_pub.get_content_instance("content_instance_slug")
-    def get_content_instance(slug)
-      @client.raw('get', "/content/content-instances/#{slug}")
-    end
-
-    ##
-    # === Get Content Pages.
-    # Get all content pages.
-    #
-    # ==== Parameters
-    # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
-    def get_content_bundles(options = nil)
-      @client.raw('get', '/content/content-bundles', options)
-    end
-
-    ##
-    # === Get Content Page.
-    # Get a single content page.
-    #
-    # ==== Parameters
-    # slug:: (String) -- It's the string identifier generated by Mints.
-    #
-    # ==== Example
-    #     @data = @mints_pub.get_content_page("test-page")
-    def get_content_page(slug, options = nil)
-      warn '[DEPRECATED] The get_content_page method is deprecated and will be removed in the future, use get_content_bundle instead'
-      @client.raw('get', "/content/content-pages/#{slug}", options)
-    end
-
-    ##
-    # === Get Content Bundle.
-    # Get a single content bundle.
-    #
-    # ==== Parameters
-    # slug:: (String) -- It's the string identifier generated by Mints.
-    #
-    # ==== Example
-    #     @data = @mints_pub.get_content_bundle("test-page")
-    def get_content_bundle(slug, options = nil)
-      @client.raw('get', "/content/content-bundles/#{slug}", options)
-    end
-
-    ### V1/ECOMMERCE ###
-
-    ##
-    # === Get Locations.
-    # Get all locations.
-    #
-    # ==== Parameters
-    # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
-    # use_post:: (Boolean) -- Variable to determine if the request is by 'post' or 'get' functions.
-    #
-    # ==== First Example
-    #     @data = @mints_pub.get_locations
-    #
-    # ==== Second Example
-    #     options = { fields: "title" }
-    #     @data = @mints_pub.get_locations(options)
-    #
-    # ==== Third Example
-    #     options = { fields: "title" }
-    #     @data = @mints_pub.get_locations(options, false)
-    def get_locations(options = nil, use_post = true)
-      get_query_results('/ecommerce/locations', options, use_post)
-    end
-
-    ##
-    # === Get Products.
-    # Get a collection of products.
-    #
-    # ==== Parameters
-    # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
-    # use_post:: (Boolean) -- Variable to determine if the request is by 'post' or 'get' functions.
-    #
-    # ==== First Example
-    #     @data = @mints_pub.get_products
-    #
-    # ==== Second Example
-    #     options = { fields: "title" }
-    #     @data = @mints_pub.get_products(options)
-    #
-    # ==== Third Example
-    #     options = { fields: "title" }
-    #     @data = @mints_pub.get_products(options, false)
-    def get_products(options = nil, use_post = true)
-      get_query_results('/ecommerce/products', options, use_post)
-    end
-
-    ##
-    # === Get Product.
-    # Get a single product.
-    #
-    # ==== Parameters
-    # slug:: (String) -- It's the string identifier generated by Mints.
-    # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
-    #
-    # ==== First Example
-    #     @data = @mints_pub.get_product("product_slug")
-    #
-    # ==== Second Example
-    #     options = {
-    #       fields: 'id, slug'
-    #     }
-    #     @data = @mints_pub.get_product("lego-set", options)
-    def get_product(slug, options = nil)
-      @client.raw('get', "/ecommerce/products/#{slug}", options)
-    end
-
-    ### V1/CONFIG ###
-
-    ##
-    # === Get Public Folders.
-    # Get a collection of public folders.
-    #
-    # ==== Parameters
-    # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
-    #
-    # ==== First Example
-    #     options = {
-    #       object_type: "products"
-    #     }
-    #     @data = @mints_pub.get_public_folders(options)
-    #
-    # ==== Second Example
-    #     options = {
-    #       object_type: "products",
-    #       fields: "id",
-    #       sort: "-id"
-    #     }
-    #     @data = @mints_pub.get_public_folders(options)
-    def get_public_folders(options = nil)
-      @client.raw('get', '/config/public-folders', options)
-    end
-
-    ##
-    # === Get Public Folder.
-    # Get a public folder info.
-    #
-    # ==== Parameters
-    # slug:: (String) -- It's the string identifier generated by Mints.
-    # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
-    #
-    # ==== First Example
-    #     options = {
-    #       object_type: "products"
-    #     }
-    #     @data = @mints_pub.get_public_folder('yellow', options)
-    #
-    # ==== Second Example
-    #     options = {
-    #       object_type: "products",
-    #       fields: 'id, title'
-    #     }
-    #     @data = @mints_pub.get_public_folder('yellow', options)
-    def get_public_folder(slug, options)
-      @client.raw('get', "/config/public-folders/#{slug}", options)
-    end
-
-    ##
-    # === Get Tags.
-    # Get a collection of tags.
-    #
-    # ==== Parameters
-    # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
-    #
-    # ==== First Example
-    #     @data = @mints_pub.get_tags
-    #
-    # ==== Second Example
-    #     options = {
-    #       fields: "id, tag"
-    #     }
-    #     @data = @mints_pub.get_tags(options)
-    def get_tags(options = nil)
-      @client.raw('get', '/config/tags', options)
-    end
-
-    ##
-    # === Get Tag.
-    # Get a single tag.
-    #
-    # ==== Parameters
-    # slug:: (String) -- It's the string identifier generated by Mints.
-    # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
-    #
-    # ==== First Example
-    #     @data = @mints_pub.get_tag("tag_slug")
-    #
-    # ==== Second Example
-    #     options = {
-    #       fields: "id, tag"
-    #     }
-    #     @data = @mints_pub.get_tag("tag-example", options)
-    def get_tag(slug, options = nil)
-      @client.raw('get', "/config/tags/#{slug}", options)
-    end
-
-    ##
-    # === Get Taxonomies.
-    # Get a collection of taxonomies.
-    #
-    # ==== Parameters
-    # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
-    # use_post:: (Boolean) -- Variable to determine if the request is by 'post' or 'get' functions.
-    #
-    # ==== First Example
-    #     @data = @mints_pub.get_taxonomies
-    #
-    # ==== Second Example
-    #     options = {
-    #       fields: 'id, title'
-    #     }
-    #     @data = @mints_pub.get_taxonomies(options)
-    #
-    # ==== Third Example
-    #     options = {
-    #       fields: 'id, title'
-    #     }
-    #     @data = @mints_pub.get_taxonomies(options, false)
-    def get_taxonomies(options = nil, use_post = true)
-      get_query_results('/config/taxonomies', options, use_post)
-    end
-
-    ##
-    # === Get Taxonomy.
-    # Get a single taxonomy.
-    #
-    # ==== Parameters
-    # slug:: (String) -- It's the string identifier generated by Mints.
-    # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
-    #
-    # ==== First Example
-    #     @data = @mints_pub.get_taxonomy('taxonomy_slug')
-    #
-    # ==== Second Example
-    #     options = {
-    #       fields: 'title'
-    #     }
-    #     @data = @mints_pub.get_taxonomy('taxonomy_slug', options)
-    def get_taxonomy(slug, options = nil)
-      @client.raw('get', "/config/taxonomies/#{slug}", options)
-    end
-
-    ##
-    # === Get Attributes.
-    # Get a collection of attributes.
-    #
-    # ==== Example
-    #     @data = @mints_pub.get_attributes
-    def get_attributes
-      @client.raw('get', '/config/attributes')
     end
 
     def send_user_magic_link(email_or_phone, template_slug, redirect_url = '', life_time = 1440, max_visits = nil, driver = 'email')
