+++ lib/pub/config/public_folders.rb
@@ -1,0 +1,51 @@
+# frozen_string_literal: true
+
+module PublicFolders
+  ##
+  # === Get Public Folders.
+  # Get a collection of public folders.
+  #
+  # ==== Parameters
+  # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
+  #
+  # ==== First Example
+  #     options = {
+  #       object_type: "products"
+  #     }
+  #     @data = @mints_pub.get_public_folders(options)
+  #
+  # ==== Second Example
+  #     options = {
+  #       object_type: "products",
+  #       fields: "id",
+  #       sort: "-id"
+  #     }
+  #     @data = @mints_pub.get_public_folders(options)
+  def get_public_folders(options = nil)
+    @client.raw('get', '/config/public-folders', options)
+  end
+
+  ##
+  # === Get Public Folder.
+  # Get a public folder info.
+  #
+  # ==== Parameters
+  # slug:: (String) -- It's the string identifier generated by Mints.
+  # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
+  #
+  # ==== First Example
+  #     options = {
+  #       object_type: "products"
+  #     }
+  #     @data = @mints_pub.get_public_folder('yellow', options)
+  #
+  # ==== Second Example
+  #     options = {
+  #       object_type: "products",
+  #       fields: 'id, title'
+  #     }
+  #     @data = @mints_pub.get_public_folder('yellow', options)
+  def get_public_folder(slug, options)
+    @client.raw('get', "/config/public-folders/#{slug}", options)
+  end
+end
