+++ lib/pub/config/tags.rb
@@ -1,0 +1,42 @@
+# frozen_string_literal: true
+
+module Tags
+  ##
+  # === Get Tags.
+  # Get a collection of tags.
+  #
+  # ==== Parameters
+  # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
+  #
+  # ==== First Example
+  #     @data = @mints_pub.get_tags
+  #
+  # ==== Second Example
+  #     options = {
+  #       fields: "id, tag"
+  #     }
+  #     @data = @mints_pub.get_tags(options)
+  def get_tags(options = nil)
+    @client.raw('get', '/config/tags', options)
+  end
+
+  ##
+  # === Get Tag.
+  # Get a single tag.
+  #
+  # ==== Parameters
+  # slug:: (String) -- It's the string identifier generated by Mints.
+  # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
+  #
+  # ==== First Example
+  #     @data = @mints_pub.get_tag("tag_slug")
+  #
+  # ==== Second Example
+  #     options = {
+  #       fields: "id, tag"
+  #     }
+  #     @data = @mints_pub.get_tag("tag-example", options)
+  def get_tag(slug, options = nil)
+    @client.raw('get', "/config/tags/#{slug}", options)
+  end
+end
