+++ lib/pub/config/taxonomies.rb
@@ -1,0 +1,49 @@
+# frozen_string_literal: true
+
+module Taxonomies
+  ##
+  # === Get Taxonomies.
+  # Get a collection of taxonomies.
+  #
+  # ==== Parameters
+  # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
+  # use_post:: (Boolean) -- Variable to determine if the request is by 'post' or 'get' functions.
+  #
+  # ==== First Example
+  #     @data = @mints_pub.get_taxonomies
+  #
+  # ==== Second Example
+  #     options = {
+  #       fields: 'id, title'
+  #     }
+  #     @data = @mints_pub.get_taxonomies(options)
+  #
+  # ==== Third Example
+  #     options = {
+  #       fields: 'id, title'
+  #     }
+  #     @data = @mints_pub.get_taxonomies(options, false)
+  def get_taxonomies(options = nil, use_post = true)
+    get_query_results('/config/taxonomies', options, use_post)
+  end
+
+  ##
+  # === Get Taxonomy.
+  # Get a single taxonomy.
+  #
+  # ==== Parameters
+  # slug:: (String) -- It's the string identifier generated by Mints.
+  # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
+  #
+  # ==== First Example
+  #     @data = @mints_pub.get_taxonomy('taxonomy_slug')
+  #
+  # ==== Second Example
+  #     options = {
+  #       fields: 'title'
+  #     }
+  #     @data = @mints_pub.get_taxonomy('taxonomy_slug', options)
+  def get_taxonomy(slug, options = nil)
+    @client.raw('get', "/config/taxonomies/#{slug}", options)
+  end
+end
