+++ lib/pub/content/assets.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Assets
+  ##
+  # === Get Asset Info.
+  # Get a description of an Asset.
+  #
+  # ==== Parameters
+  # slug:: (String) -- It's the string identifier of the asset.
+  #
+  # ==== Example
+  #     @data = @mints_pub.get_asset_info("asset_slug")
+  def get_asset_info(slug)
+    @client.raw('get', "/content/asset-info/#{slug}")
+  end
+end
