+++ lib/pub/content/story_versions.rb
@@ -1,0 +1,46 @@
+# frozen_string_literal: true
+
+module StoryVersions
+  ##
+  # === Get Story Versions.
+  # Get a collection of story version.
+  #
+  # ==== Parameters
+  # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
+  # use_post:: (Boolean) -- Variable to determine if the request is by 'post' or 'get' functions.
+  #
+  # ==== First Example
+  #     @data = @mints_pub.get_story_versions
+  #
+  # ==== Second Example
+  #     options = {
+  #       fields: 'id, title'
+  #     }
+  #     @data = @mints_pub.get_story_versions(options)
+  #
+  # ==== Third Example
+  #     options = {
+  #       fields: 'id, title'
+  #     }
+  #     @data = @mints_pub.get_story_versions(options, false)
+  def get_story_versions(options = nil, use_post = true)
+    get_query_results('/content/story-versions', options, use_post)
+  end
+
+  ##
+  # === Get Story Version.
+  # Get a single story version.
+  #
+  # ==== Parameters
+  # slug:: (String) -- It's the string identifier generated by Mints.
+  # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
+  #
+  # ==== First Example
+  #     @data = @mints_pub.get_story_version('story_slug')
+  #
+  # ==== Second Example
+  #     @data = @mints_pub.get_story_version('story_slug', options)
+  def get_story_version(slug, options = nil)
+    @client.raw('get', "/content/story-versions/#{slug}", options)
+  end
+end
