+++ lib/pub/ecommerce/ecommerce.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+### V1/ECOMMERCE ###
+
+require_relative './locations'
+require_relative './products'
+require_relative './orders'
+
+module Ecommerce
+  include Locations
+  include Products
+  include Orders
+
+end
