+++ lib/pub/ecommerce/products.rb
@@ -1,0 +1,45 @@
+# frozen_string_literal: true
+
+module Products
+  ##
+  # === Get Products.
+  # Get a collection of products.
+  #
+  # ==== Parameters
+  # options:: (Hash) -- List of {Resource collection Options}[#class-Mints::Pub-label-Resource+collections+options+] shown above can be used as parameter.
+  # use_post:: (Boolean) -- Variable to determine if the request is by 'post' or 'get' functions.
+  #
+  # ==== First Example
+  #     @data = @mints_pub.get_products
+  #
+  # ==== Second Example
+  #     options = { fields: "title" }
+  #     @data = @mints_pub.get_products(options)
+  #
+  # ==== Third Example
+  #     options = { fields: "title" }
+  #     @data = @mints_pub.get_products(options, false)
+  def get_products(options = nil, use_post = true)
+    get_query_results('/ecommerce/products', options, use_post)
+  end
+
+  ##
+  # === Get Product.
+  # Get a single product.
+  #
+  # ==== Parameters
+  # slug:: (String) -- It's the string identifier generated by Mints.
+  # options:: (Hash) -- List of {Single Resource Options}[#class-Mints::Pub-label-Single+resource+options] shown above can be used as parameter.
+  #
+  # ==== First Example
+  #     @data = @mints_pub.get_product("product_slug")
+  #
+  # ==== Second Example
+  #     options = {
+  #       fields: 'id, slug'
+  #     }
+  #     @data = @mints_pub.get_product("lego-set", options)
+  def get_product(slug, options = nil)
+    @client.raw('get', "/ecommerce/products/#{slug}", options)
+  end
+end
